#21 Moving homekit classes do different files
