+++ gadget_publishers/homekit/homekit_gadget_update_interface.py
@@ -1,0 +1,16 @@
+from abc import abstractmethod, ABCMeta
+
+
+class GadgetPublisherHomekitInterface(metaclass=ABCMeta):
+    """Interface for the gadget update forwarding"""
+
+    @abstractmethod
+    def receive_update(self, gadget_name: str, update_data: dict) -> None:
+        """
+        Receive an update for the named gadget
+
+        :param gadget_name: Name of the updated gadget
+        :param update_data: Information about the updatd characteristics
+        :return: None
+        """
+        pass
